feat: aggregation method config
