+++ wormhole/aggregator/src/lib.rs
@@ -1,5 +1,5 @@
 pub mod aggregator;
-pub mod circuit;
+pub mod circuits;
 
 /// The maximum numbers of proofs to aggregate into a composite proof.
 pub const DEFAULT_NUM_PROOFS_TO_AGGREGATE: usize = 10;
