+++ wormhole/tests/src/aggregator/aggregator_tests.rs
@@ -1,5 +1,8 @@
 #![cfg(test)]
-use wormhole_aggregator::{aggregator::WormholeProofAggregator, DEFAULT_NUM_PROOFS_TO_AGGREGATE};
+use wormhole_aggregator::{
+    aggregator::{AggregationMethod, WormholeProofAggregator},
+    DEFAULT_NUM_PROOFS_TO_AGGREGATE,
+};
 use wormhole_circuit::inputs::CircuitInputs;
 use wormhole_prover::WormholeProver;
 
@@ -13,8 +16,10 @@ fn push_proof_to_buffer() {
     let inputs = CircuitInputs::test_inputs();
     let proof = prover.commit(&inputs).unwrap().prove().unwrap();
 
-    let mut aggregator =
-        WormholeProofAggregator::<{ DEFAULT_NUM_PROOFS_TO_AGGREGATE }>::new(circuit_config());
+    let mut aggregator = WormholeProofAggregator::<{ DEFAULT_NUM_PROOFS_TO_AGGREGATE }>::new(
+        circuit_config(),
+        AggregationMethod::Flat,
+    );
     aggregator.push_proof(proof).unwrap();
 
     let proofs_buffer = aggregator.proofs_buffer.unwrap();
@@ -28,8 +33,10 @@ fn push_proof_to_full_buffer() {
     let inputs = CircuitInputs::test_inputs();
     let proof = prover.commit(&inputs).unwrap().prove().unwrap();
 
-    let mut aggregator =
-        WormholeProofAggregator::<{ DEFAULT_NUM_PROOFS_TO_AGGREGATE }>::new(circuit_config());
+    let mut aggregator = WormholeProofAggregator::<{ DEFAULT_NUM_PROOFS_TO_AGGREGATE }>::new(
+        circuit_config(),
+        AggregationMethod::Flat,
+    );
 
     // Fill up the proof buffer.
     for _ in 0..DEFAULT_NUM_PROOFS_TO_AGGREGATE {
@@ -50,8 +57,10 @@ fn aggregate_single_proof() {
     let inputs = CircuitInputs::test_inputs();
     let proof = prover.commit(&inputs).unwrap().prove().unwrap();
 
-    let mut aggregator =
-        WormholeProofAggregator::<{ DEFAULT_NUM_PROOFS_TO_AGGREGATE }>::new(circuit_config());
+    let mut aggregator = WormholeProofAggregator::<{ DEFAULT_NUM_PROOFS_TO_AGGREGATE }>::new(
+        circuit_config(),
+        AggregationMethod::Flat,
+    );
     aggregator.push_proof(proof).unwrap();
 
     aggregator.aggregate().unwrap();
